refactor(core): relocate SystemThemeListener and RelativeTimeTicker
Audit fix Batch D — Section 3 of the v2.0.0 design catalogs these as
controllers, so they belong under `packages/core/src/controllers/`
not in `theme.ts` / `auto-save.ts`. The relocation aligns the layer
table in `.claude/rules/packages/core.md` with the actual file
structure.

- packages/core/src/controllers/systemThemeListener.ts hosts
  createVizelSystemThemeListener (moved from theme.ts).
- packages/core/src/controllers/relativeTimeTicker.ts hosts
  createVizelRelativeTimeTicker (moved from auto-save.ts) and imports
  `formatVizelRelativeTime` from auto-save for the actual formatting.
- controllers/index.ts re-exports the new entries.
- packages/core/src/index.ts moves the two factories and their
  options/return types from the "Theme" / "Auto-save" sections into
  the "Controllers" section.
